+++ go.mod
@@ -3,7 +3,6 @@ module github.com/ajski1701/go-meater-meter
 go 1.16
 
 require (
-	github.com/influxdata/influxdb-client-go v1.4.0
 	github.com/influxdata/influxdb-client-go/v2 v2.9.0
 	gopkg.in/ini.v1 v1.66.6
 )
